Added INSUFFICIENT_RESOURCES reply
There currently is no standardized reply when no physical nodes available, no pyshical nodes with the requested number of interfaces available or it is not possible to create an additional VM.
A new reply code would be useful for these cases, where the requested resources could not be allocated.

This reply code would enable clients to easily detect this case. That would allow jFed to inform the user of the reason for the failure, and would allow "fedmon", the fed4fire FLS monitor, to flag these failures as a login test warning instead of a failure.

(Feedback is needed about the name of the error code, the value, and the exact cases targeted by this error. The possible overlap with INSUFFICIENT_BANDWIDTH should also be handled.)